+++ src/main/java/io/aqsha/api/repositories/UserRepository.java
@@ -2,8 +2,8 @@ package io.aqsha.api.repositories;
 
 import org.springframework.data.repository.CrudRepository;
 
-import io.aqsha.api.models.User;;
+import io.aqsha.api.models.User;
 
-public interface UserRepository extends CrudRepository<User, Integer> {
+public interface UserRepository extends CrudRepository<User, Long> {
 
 }
